Jacobi.jl documentation

Jacobi theta functions and related functions.

Jacobi.jl documentation

Functions

Index

Functions

Jacobi.ljtheta1 — Function

ljtheta1(z, tau)

Logarithm of the first Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.jtheta1 — Function

jtheta1(z, tau)

First Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.ljtheta2 — Function

ljtheta2(z, tau)

Logarithm of the second Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.jtheta2 — Function

jtheta2(z, tau)

Second Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.ljtheta3 — Function

ljtheta3(z, tau)

Logarithm of the third Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.jtheta3 — Function

jtheta3(z, tau)

Third Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.ljtheta4 — Function

ljtheta4(z, tau)

Logarithm of the fourth Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.jtheta4 — Function

jtheta4(z, tau)

Fourth Jacobi theta function.

Arguments

z: complex number or vector of complex numbers

tau: complex number with nonnegative imaginary part

Jacobi.jtheta1dash — Function

jtheta1dash(z, tau)

Derivative of the first Jacobi theta function.

Arguments

z: complex number

tau: complex number with nonnegative imaginary part

Jacobi.etaDedekind — Function

etaDedekind(tau)

Dedekind eta function.

Arguments

tau: complex number with nonnegative imaginary part

Jacobi.lambda — Function

lambda(tau)

Lambda modular function.

Arguments

tau: complex number with nonnegative imaginary part

Jacobi.kleinj — Function

kleinj(tau)

Klein j-invariant function.

Arguments

tau: complex number with nonnegative imaginary part

Jacobi.CarlsonRF — Function

CarlsonRF(x, y, z)

Carlson 'RF' integral.

Arguments

x,y,z: complex numbers; at most one of them can be zero

Jacobi.CarlsonRD — Function

CarlsonRD(x, y, z)

Carlson 'RD' integral.

Arguments

x,y,z: complex numbers; at most one of them can be zero

Jacobi.ellipticF — Function

ellipticF(phi, m)

Incomplete elliptic integral of the first kind.

Arguments

phi: complex number, the amplitude

m: complex number, the squared modulus

Jacobi.ellipticK — Function

ellipticK(m)

Complete elliptic integral of the first kind.

Arguments

m: complex number, the squared modulus

Jacobi.ellipticE — Function

ellipticE(phi, m)

Incomplete elliptic integral of the second kind.

Arguments

phi: complex number, the amplitude

m: complex number, the squared modulus

ellipticE(m)

Complete elliptic integral of the second kind.

Arguments

m: complex number, the squared modulus

Jacobi.agm — Function

agm(x, y)

Arithmetic-geometric mean.

Arguments

x,y: complex numbers

Jacobi.EisensteinE2 — Function

EisensteinE2(q)

Eisenstein E-series of weight 2.

Arguments

q: nome, complex number; it must not be a negative real number and its modulus must be strictly smaller than 1

Jacobi.wp — Function

wp(z; tau, omega)

Weierstrass p-function.

Arguments

z: complex number or vector of complex numbers

tau: half-periods ratio, complex number with non negative imaginary part

omega: half-periods, a pair of complex numbers; exactly one of tau or omega must be given

Jacobi.wsigma — Function

wsigma(z; tau, omega)

Weierstrass sigma-function.

Arguments

z: complex number or vector of complex numbers

tau: half-periods ratio, complex number with non negative imaginary part

omega: half-periods, a pair of complex numbers; exactly one of tau or omega must be given

Jacobi.wzeta — Function

wzeta(z; tau, omega)

Weierstrass zeta-function.

Arguments

z: complex number or vector of complex numbers

tau: half-periods ratio, complex number with non negative imaginary part

omega: half-periods, a pair of complex numbers; exactly one of tau or omega must be given

Index

Jacobi.CarlsonRD

Jacobi.CarlsonRF

Jacobi.EisensteinE2

Jacobi.agm

Jacobi.ellipticE

Jacobi.ellipticF

Jacobi.ellipticK

Jacobi.etaDedekind

Jacobi.jtheta1

Jacobi.jtheta1dash

Jacobi.jtheta2

Jacobi.jtheta3

Jacobi.jtheta4

Jacobi.kleinj

Jacobi.lambda

Jacobi.ljtheta1

Jacobi.ljtheta2

Jacobi.ljtheta3

Jacobi.ljtheta4

Jacobi.wp

Jacobi.wsigma

Jacobi.wzeta